starting network restructure
